Separated data structures to their own files- saving works for single and mean scores. Bracket saving (relative remains unchanged)
