+++ Masters-Unity/Assets/Scripts/BracketData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 14c7c26f143a9a545b32f1332aec1446
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
