+++ Masters-Unity/Assets/Scripts/ComfortData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7f79473bc5ef25448dc4b9dde7fd62c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
